Moved logic from "aggregation chart service" into "chart service"
