+++ CSETWebNg/src/app/aggregation/compare-analytics/compare-bestworst/compare-bestworst.component.ts
@@ -24,7 +24,7 @@
 import { Component, OnInit } from '@angular/core';
 import { NavigationAggregService } from '../../../services/navigationAggreg.service';
 import { AggregationService } from '../../../services/aggregation.service';
-import { AggregationChartService } from '../../../services/aggregation-chart.service';
+import { ChartService } from '../../../services/chart.service';
 import  Chart  from 'chart.js/auto';
 
 @Component({
@@ -42,7 +42,7 @@ export class CompareBestworstComponent implements OnInit {
 
   constructor(
     public aggregationSvc: AggregationService,
-    public aggregChartSvc: AggregationChartService
+    public chartSvc: ChartService
   ) { }
 
   ngOnInit() {
@@ -101,6 +101,6 @@ export class CompareBestworstComponent implements OnInit {
     if (!!this.chartAnswerBreakdown) {
       this.chartAnswerBreakdown.destroy();
     }
-    this.chartAnswerBreakdown = this.aggregChartSvc.buildStackedHorizBarChart('canvasAnswerBreakdown', x);
+    this.chartAnswerBreakdown = this.chartSvc.buildStackedHorizBarChart('canvasAnswerBreakdown', x);
   }
 }
