+++ CSETWebNg/src/app/aggregation/compare-analytics/compare-summary/compare-summary.component.ts
@@ -23,7 +23,6 @@
 ////////////////////////////////
 import { Component, OnInit } from '@angular/core';
 import { AggregationService } from '../../../services/aggregation.service';
-import { AggregationChartService } from '../../../services/aggregation-chart.service';
 import { Chart } from 'chart.js'
 import { ChartService } from '../../../services/chart.service';
 
@@ -43,7 +42,6 @@ export class CompareSummaryComponent implements OnInit {
 
   constructor(
     public aggregationSvc: AggregationService,
-    public aggregChartSvc: AggregationChartService,
     public chartSvc: ChartService
   ) { }
 
@@ -63,7 +61,7 @@ export class CompareSummaryComponent implements OnInit {
         ds.borderColor = '#004c75';
       });
 
-      this.chartOverallAverage = this.aggregChartSvc.buildHorizBarChart('canvasOverallAverage', x, false);
+      this.chartOverallAverage = this.chartSvc.buildHorizBarChart('canvasOverallAverage', x, false, true);
     });
 
 
@@ -96,10 +94,10 @@ export class CompareSummaryComponent implements OnInit {
       }
       x.options.maintainAspectRatio = false;   
       
-      this.chartCategoryAverage = this.aggregChartSvc.buildHorizBarChart('canvasCategoryAverage', x, false);
+      this.chartCategoryAverage = this.chartSvc.buildHorizBarChart('canvasCategoryAverage', x, false, true);
       
       if (this.chartCategoryAverage.canvas) {
-        (<HTMLElement>this.chartCategoryAverage.canvas.parentNode).style.height = this.aggregChartSvc.calcHbcHeightPixels(x);
+        (<HTMLElement>this.chartCategoryAverage.canvas.parentNode).style.height = this.chartSvc.calcHbcHeightPixels(x);
       }
     });
   }
