+++ CSETWebNg/src/app/services/chart.service.ts
@@ -24,9 +24,9 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ConfigService } from './config.service';
-import Chart from 'chart.js/auto';
 import { Utilities } from './utilities.service';
-import { Options } from 'selenium-webdriver';
+import Chart from 'chart.js/auto';
+
 
 /**
  * The eventual home for one-stop shopping for the various
@@ -41,6 +41,140 @@ export class ChartService {
     private http: HttpClient,
     private configSvc: ConfigService
   ) { }
+
+  /**
+  * Builds a line chart from the aggregation API response.
+  * @param canvasId
+  * @param x
+  */
+  buildLineChart(canvasId: string, x: any) {
+    const chartColors = new ChartColors();
+
+    // add display characteristics
+    for (let i = 0; i < x.datasets.length; i++) {
+      const ds = x.datasets[i];
+      ds.borderColor = chartColors.getLineColor(i);
+      ds.backgroundColor = ds.borderColor;
+      ds.pointRadius = 8;
+      ds.lineTension = 0;
+      ds.fill = false;
+    }
+    let tempChart = Chart.getChart(canvasId);
+    if (tempChart) {
+      tempChart.destroy();
+    }
+    return new Chart(canvasId, {
+      type: 'line',
+      data: {
+        labels: x.labels,
+        datasets: x.datasets
+      },
+      options: {
+        maintainAspectRatio: true,
+        plugins: {
+          legend: { position: 'left' },
+          tooltip: {
+            callbacks: {
+              label: ((context) =>
+                context.dataset.label + (!!context.dataset.label ? ': ' : ' ')
+                + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%')
+            }
+          }
+        }
+      }
+    });
+  }
+
+  /**
+  *
+  * @param canvasId
+  * @param x
+  * @param showLegend
+  */
+  buildBarChart(canvasId: string, x: any, showLegend: boolean) {
+    if (!x.labels) {
+      x.labels = [];
+    }
+    x.datasets.forEach(ds => {
+      if (!ds.label) {
+        ds.label = '';
+      }
+    });
+    let tempChart = Chart.getChart(canvasId);
+    if (tempChart) {
+      tempChart.destroy();
+    }
+    return new Chart(canvasId, {
+      type: 'bar',
+      data: {
+        labels: x.labels,
+        datasets: x.datasets
+      },
+      options: {
+        maintainAspectRatio: true,
+        responsive: false,
+        plugins: {
+          legend: { display: showLegend, position: 'top' }
+        }
+      }
+    });
+  }
+
+  /**
+ * Builds a horizontal bar chart.  The x-axis and tooltips are always formatted as %
+ * @param canvasId
+ * @param x
+ */
+  buildHorizBarChart(canvasId: string, x: any, showLegend: boolean, zeroHundred: boolean) {
+    if (!x.labels) {
+      x.labels = [];
+    }
+    x.datasets.forEach(ds => {
+      if (!ds.label) {
+        ds.label = '';
+      }
+    });
+
+    let maintainAspectRatio = true;
+    if (x.hasOwnProperty('options') && x.options.hasOwnProperty('maintainAspectRatio')) {
+      maintainAspectRatio = x.options.maintainAspectRatio;
+    }
+    let tempChart = Chart.getChart(canvasId);
+    if (tempChart) {
+      tempChart.destroy();
+    }
+
+    var myOptions: any = {
+      indexAxis: 'y',
+      maintainAspectRatio: maintainAspectRatio,
+      responsive: true,
+      plugins: {
+        legend: { display: showLegend, position: 'top' },
+        tooltip: {
+          callbacks: {
+            label: ((context) =>
+              context.dataset.label + (!!context.dataset.label ? ': ' : ' ')
+              + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%')
+          }
+        }
+      }
+    };
+
+    // set the scale if desired
+    if (zeroHundred) {
+      myOptions.scale = { min: 0, max: 100 };
+    }
+
+
+    return new Chart(canvasId, {
+      type: 'bar',
+      data: {
+        labels: x.labels,
+        datasets: x.datasets
+      },
+      options: myOptions
+    });
+  }
 
   /**
    *
@@ -86,7 +220,7 @@ export class ChartService {
           tooltip: {
             callbacks: {
               label: function (context) {
-                const label = context.label + (!!context.label ? ': '  : ' ')
+                const label = context.label + (!!context.label ? ': ' : ' ')
                   + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%';
                 return label;
               }
@@ -144,7 +278,7 @@ export class ChartService {
    * return false, otherwise return true.
    */
   looksLikeAnswerDistribution(labels): boolean {
-    var answerLabels = [ 'Y', 'N', 'U', 'NA', 'A', 'I'];
+    var answerLabels = ['Y', 'N', 'U', 'NA', 'A', 'I'];
 
     for (const element of labels) {
       if (answerLabels.indexOf(element) < 0) {
@@ -156,58 +290,86 @@ export class ChartService {
   }
 
   /**
-   * Builds a horizontal bar chart.  The x-axis and tooltips are always formatted as %
+   * Builds a horizontal bar chart from the Dashboard API response.
    * @param canvasId
    * @param x
    */
-  buildHorizBarChart(canvasId: string, x: any, showLegend: boolean, zeroHundred: boolean) {
-    if (!x.labels) {
-      x.labels = [];
-    }
-    x.datasets.forEach(ds => {
-      if (!ds.label) {
-        ds.label = '';
-      }
-    });
+  buildCategoryPercentChart(canvasId: string, x: any) {
+    const chartColors = new ChartColors();
 
-    let maintainAspectRatio = true;
-    if (x.hasOwnProperty('options') && x.options.hasOwnProperty('maintainAspectRatio')) {
-      maintainAspectRatio = x.options.maintainAspectRatio;
+    for (let i = 0; i < x.datasets.length; i++) {
+      const ds = x.datasets[i];
+      if (ds.label === '') {
+        ds.label = 'A';
+      }
+      ds.borderColor = chartColors.getNextBarColor();
+      ds.backgroundColor = ds.borderColor;
     }
+
     let tempChart = Chart.getChart(canvasId);
     if (tempChart) {
       tempChart.destroy();
     }
-
-    var myOptions: any = {
-      indexAxis: 'y',
-      maintainAspectRatio: maintainAspectRatio,
-      responsive: true,
-      plugins: {
-        legend: { display: showLegend, position: 'top' },
-        tooltip: {
-          callbacks: {
-            label: ((context) =>
-              context.dataset.label + (!!context.dataset.label ? ': '  : ' ')
-              + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%')
-          }
-        }
-      }
-    };
-
-    // set the scale if desired
-    if (zeroHundred) {
-      myOptions.scale = { min: 0, max: 100 };
-    }
-
-
     return new Chart(canvasId, {
       type: 'bar',
       data: {
         labels: x.labels,
         datasets: x.datasets
       },
-      options: myOptions
+
+      options: {
+        indexAxis: 'y',
+        maintainAspectRatio: false,
+        plugins: {
+          tooltip: {
+            callbacks: {
+              label: ((context) =>
+                context.dataset.label + (!!context.dataset.label ? ': ' : ' ')
+                + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%')
+            }
+          }
+        }
+
+      }
+    });
+  }
+
+  /**
+   *
+   */
+  buildStackedHorizBarChart(canvasId: string, x: any) {
+    let tempChart = Chart.getChart(canvasId);
+    if (tempChart) {
+      tempChart.destroy();
+    }
+    return new Chart(canvasId, {
+      type: 'bar',
+      data: {
+        labels: x.labels,
+        datasets: x.datasets
+      },
+      options: {
+        indexAxis: 'y',
+        animation: { duration: 100 }, // general animation time
+        scales: {
+          x: {
+            stacked: true
+          },
+          y: {
+            stacked: true
+          }
+        },
+        maintainAspectRatio: false,
+        plugins: {
+          tooltip: {
+            callbacks: {
+              label: ((context) =>
+                context.dataset.label + (!!context.dataset.label ? ': ' : ' ')
+                + (<Number>context.dataset.data[context.dataIndex]).toFixed() + '%')
+            }
+          }
+        }
+      }
     });
   }
 
@@ -236,5 +398,111 @@ export class ChartService {
       default:
         return '#000000';
     }
+  }
+
+  /**
+  * Calculates a good height for a horizontal bar chart
+  * based on the number of datasets and data items,
+  * i.e., the number of horizontal bars to accommodate.
+  * @param x
+  */
+  calcHbcHeightPixels(x): string {
+    // calculate the number of bars in the graph
+    let maxDatasetLength = x.datasets[0].data.length;
+    // calculate a good height for the chart's container
+    let h = maxDatasetLength * x.datasets.length * 20;
+    return (h + 50) + "px";
+  }
+}
+
+
+/**
+ * A service that supplies colors for charts.
+ */
+@Injectable()
+export class ChartColors {
+  /**
+   * These colors are used for bar charts with multiple assessments.
+   */
+  barColorSequence = [
+    '#0000FF',
+    '#FFD700',
+    '#008000',
+    '#6495ED',
+    '#006400',
+    '#F0E68C',
+    '#00008B',
+    '#008B8B',
+    '#FFFACD',
+    '#483D8B',
+    '#2F4F4F',
+    '#9400D3',
+    '#1E90FF',
+    '#B22222',
+    '#FFFAF0',
+    '#228B22',
+    '#DAA520',
+    '#ADFF2F',
+    '#4B0082',
+    '#000080'
+  ];
+  nextBarSequence: number = -1;
+
+  /**
+   * These colors are used for the Overall Comparison chart
+   */
+  bluesColorSequence = [
+    '#B0BFDB',
+    '#7E97C2',
+    '#4180CB',
+    '#386FB3',
+    '#295588'
+  ];
+  nextBluesSequence: number = -1;
+
+  /**
+   * These colors are used for line charts.
+   */
+  lineColors = [
+    '#3e7bc4',
+    '#81633f',
+    '#9ac04a',
+    '#7c5aa6',
+    '#38adcc'
+  ];
+
+
+  /**
+   * Returns the next color in the sequence.
+   * Wraps around when exhausted.
+   */
+  getNextBarColor() {
+    if (this.nextBarSequence > this.barColorSequence.length - 1) {
+      this.nextBarSequence = -1;
+    }
+
+    return this.barColorSequence[++this.nextBarSequence];
+  }
+
+
+  /**
+   * Returns the next color in the sequence.
+   * Wraps around when exhausted.
+   */
+  getNextBluesBarColor() {
+    if (this.nextBluesSequence > this.bluesColorSequence.length - 1) {
+      this.nextBluesSequence = -1;
+    }
+
+    return this.bluesColorSequence[++this.nextBluesSequence];
+  }
+
+
+  /**
+   * Returns the specified line color based on position.
+   * @param idx
+   */
+  getLineColor(i: number): string {
+    return this.lineColors[i];
   }
 }
